Slight Code Modularisation
