+++ gui.py
@@ -2,8 +2,8 @@ import socket
 import tkinter as tk
 import threading
 from tkinter import filedialog
-from tkinter import messagebox  # Import messagebox for showing alerts
 from networking import receive_messages, send_text
+from utils import save_file, FindDialog
 
 class CollaborativeTextEditor(tk.Tk):
     def __init__(self):
@@ -43,12 +43,7 @@ class CollaborativeTextEditor(tk.Tk):
         send_text(self.text_widget, self.client)
 
     def save_file(self):
-        file_path = filedialog.asksaveasfilename(defaultextension=".txt")
-        if file_path:
-            text_content = self.text_widget.get("1.0", "end-1c")
-            with open(file_path, "w") as file:
-                file.write(text_content)
-            messagebox.showinfo("Info", "File saved successfully.")  # Show a success message
+        save_file(self.text_widget)
 
     def cut_text(self):
         self.clipboard_clear()
@@ -72,33 +67,3 @@ class CollaborativeTextEditor(tk.Tk):
     def destroy(self):
         self.client.send('exit'.encode('utf-8'))
         super().destroy()
-
-class FindDialog(tk.Toplevel):
-    def __init__(self, parent):
-        super().__init__(parent)
-
-        self.parent = parent
-        self.title("Find")
-        self.geometry("300x100")
-
-        self.find_label = tk.Label(self, text="Find:")
-        self.find_label.pack()
-
-        self.find_entry = tk.Entry(self)
-        self.find_entry.pack()
-
-        self.find_button = tk.Button(self, text="Find", command=self.find)
-        self.find_button.pack()
-
-    def find(self):
-        search_str = self.find_entry.get()
-        text_widget = self.parent.text_widget
-        start_index = text_widget.search(search_str, "1.0", stopindex=tk.END)
-        if start_index:
-            end_index = f"{start_index}+{len(search_str)}c"
-            text_widget.tag_remove("found", "1.0", tk.END)
-            text_widget.tag_add("found", start_index, end_index)
-            text_widget.see(start_index)
-            text_widget.focus_set()
-        else:
-            messagebox.showinfo("Info", "Text not found.")
